backend/requirements.txt
fastapi
uvicorn
qrcode
python-multipart
pyzbar
Pillow
